requirements.txt
black==22.6.0
click==8.1.3
dataclasses-json==0.5.7
Faker==13.15.0
kafka-python==2.0.2
marshmallow==3.17.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typing-inspect==0.7.1
typing_extensions==4.3.0
